Transform User into UserEntity as User

+++ backend/eshop/src/main/java/com/example/eshop/repository/interfaces/OrderJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.eshop.model.Order;
 
 @Repository
-public interface OrderJpaRepository extends JpaRepository<Order, Integer> {
+public interface OrderJpaRepository extends JpaRepository<Order, Long> {
     
 }
